Pushed update

+++ lib/models/CheckInGuestReq.dart
@@ -9,11 +9,12 @@ class CheckInGuestReq {
   @JsonKey(name: 'GuestName')
   String? guestName;
   @JsonKey(name: 'CheckInDate')
-  double? checkInDate;
+  int? checkInDate;
   @JsonKey(name: 'CheckoutDate')
   int? checkoutDate;
 
-  CheckInGuestReq(this.roomId, this.guestName, this.checkInDate, this.checkoutDate);
+  CheckInGuestReq(
+      this.roomId, this.guestName, this.checkInDate, this.checkoutDate);
 
   static List<CheckInGuestReq> fromArray(List<dynamic> list) {
     List<CheckInGuestReq> result = [];
@@ -23,7 +24,8 @@ class CheckInGuestReq {
     return result;
   }
 
-  factory CheckInGuestReq.fromJson(Map<String, dynamic> json) => _$CheckInGuestReqFromJson(json);
+  factory CheckInGuestReq.fromJson(Map<String, dynamic> json) =>
+      _$CheckInGuestReqFromJson(json);
 
   Map<String, dynamic> toJson() => _$CheckInGuestReqToJson(this);
 }
